refactor drag methods to separate file
