+++ src/DashboardUI/src/components/map/CustomDirectSelectMode/CustomDirectSelectMode.tsx
@@ -2,16 +2,13 @@ import MapboxDraw, { DrawCustomMode, DrawCustomModeThis } from '@mapbox/mapbox-g
 import center from '@turf/center';
 import { LngLat } from 'mapbox-gl';
 import { Feature, GeoJSON, GeoJsonProperties, Geometry, Position } from 'geojson';
-import { generateCircleWithRadiusFromCenterPointToEdgePoint } from '../../../utilities/geometryHelpers';
-import bboxPolygon from '@turf/bbox-polygon';
-import bbox from '@turf/bbox';
-import distance from '@turf/distance';
+import { dragFeature, dragVertex } from './CustomDirectSelectModeDragHelpers';
 
 // base code used for reference: https://github.com/mapbox/mapbox-gl-draw/blob/main/src/modes/direct_select.js
 
-const baseMode = MapboxDraw.modes.direct_select;
+export const directSelectBaseMode = MapboxDraw.modes.direct_select;
 
-interface CustomDirectSelectModeState {
+export interface CustomDirectSelectModeState {
   customState: {
     isCircle: boolean;
     circleState: {
@@ -36,16 +33,16 @@ const defaultCustomState: CustomDirectSelectModeState['customState'] = {
   isRectangle: false,
 };
 
-type DrawModeInstance = DrawCustomMode & DrawCustomModeThis;
+export type DirectSelectDrawModeInstance = DrawCustomMode & DrawCustomModeThis;
 
 // DrawCustomMode type definition doesn't allow for extending the type with additional properties
 // although the js implementation of `direct_select` does exactly that
 export const CustomDirectSelectMode: DrawCustomMode = {
-  ...baseMode,
+  ...directSelectBaseMode,
 
   // when a polygon is clicked on, this method is called to initialize the state
   onSetup: function (opts): CustomDirectSelectModeState {
-    const state: CustomDirectSelectModeState = baseMode.onSetup?.call(this, opts);
+    const state: CustomDirectSelectModeState = directSelectBaseMode.onSetup?.call(this, opts);
     state.customState = defaultCustomState;
 
     if (state.feature?.properties?.isCircle) {
@@ -136,92 +133,7 @@ export const CustomDirectSelectMode: DrawCustomMode = {
       corners.map((position) => buildVertex(geojson.properties!.id, position)).forEach((vertex) => display(vertex));
     } else {
       // call base implementation
-      baseMode.toDisplayFeatures?.call(this, state, geojson, display);
+      directSelectBaseMode.toDisplayFeatures?.call(this, state, geojson, display);
     }
   },
-};
-
-const dragFeature = (_this: DrawModeInstance, state: CustomDirectSelectModeState, e: MapboxDraw.MapMouseEvent) => {
-  // call base implementation, let it handle the drag
-  baseMode.onDrag?.call(_this, state, e);
-
-  if (state.feature?.properties?.isCircle) {
-    // afterwards update the center point of the circle
-    state.customState.circleState.circleCenterPointLngLat = center(state.feature).geometry.coordinates as [
-      number,
-      number,
-    ];
-  }
-};
-
-const dragVertex = (_this: DrawModeInstance, state: CustomDirectSelectModeState, e: MapboxDraw.MapMouseEvent) => {
-  // determine whether to override the drag behavior
-  if (state.feature?.properties?.isCircle) {
-    handleDragCircleVertex(state, e);
-  } else if (state.feature?.properties?.isRectangle) {
-    handleDragRectangleVertex(state, e);
-  } else {
-    baseMode.onDrag?.call(_this, state, e);
-  }
-};
-
-const handleDragCircleVertex = (state: CustomDirectSelectModeState, e: MapboxDraw.MapMouseEvent) => {
-  // find new coordinate location via mouse position
-  const newCoordPosition = e.lngLat;
-
-  // generate new coordinates for the whole circle
-  const newCoordinates = generateCircleWithRadiusFromCenterPointToEdgePoint(
-    state.customState.circleState.circleCenterPointLngLat!,
-    newCoordPosition.toArray(),
-  );
-
-  // overwrite the circles coordinates with the new coordinates
-  state.feature?.setCoordinates(newCoordinates.geometry.coordinates);
-};
-
-const handleDragRectangleVertex = (state: CustomDirectSelectModeState, e: MapboxDraw.MapMouseEvent) => {
-  const findClosestPointToPoint = (newPoint: Position, points: Position[]): Position => {
-    return points
-      .map((point) => ({
-        point,
-        distance: distance(newPoint, point, { units: 'kilometers' }),
-      }))
-      .reduce((prevPoint, currentPoint) => (prevPoint.distance < currentPoint.distance ? prevPoint : currentPoint))
-      .point;
-  };
-
-  const computeNewRectangleCoordinates = (activeCorner: Position, oppositeCorner: Position): Position[] => {
-    const [lng1, lat1] = activeCorner;
-    const [lng2, lat2] = oppositeCorner;
-
-    return [
-      [lng1, lat1],
-      [lng2, lat1],
-      [lng2, lat2],
-      [lng1, lat2],
-    ];
-  };
-
-  const rectangle = state.feature!;
-  const corners = rectangle.getCoordinates()[0];
-  const activeCorner = findClosestPointToPoint(e.lngLat.toArray(), corners);
-  const oppositeCornerIndex = (corners.indexOf(activeCorner) + 2) % 4;
-  const oppositeCorner = corners[oppositeCornerIndex];
-
-  activeCorner[0] = e.lngLat.lng;
-  activeCorner[1] = e.lngLat.lat;
-
-  const newRectangleCoordinates = computeNewRectangleCoordinates(activeCorner, oppositeCorner);
-
-  const newRectangleGeometry = bboxPolygon(
-    bbox({
-      type: 'Feature',
-      properties: {},
-      geometry: {
-        type: 'Polygon',
-        coordinates: [newRectangleCoordinates],
-      },
-    }),
-  );
-  rectangle.setCoordinates(newRectangleGeometry.geometry.coordinates);
 };
